analytics/requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
psycopg2-binary>=2.9
mlxtend>=0.23
jupyter>=1.0
python-dotenv>=1.0
